+++ app/Actions/Auth/LogoutAction.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Actions;
+namespace App\Actions\Auth;
 
 use Tymon\JWTAuth\Facades\JWTAuth;
