+++ app/Http/Controllers/Auth/AuthController.php
@@ -1,23 +1,18 @@
 <?php
 
-namespace App\Http\Controllers\Api;
+namespace App\Http\Controllers\Auth;
 
-use App\Actions\LoginAction;
-use App\Actions\LogoutAction;
+use App\Actions\Auth\LoginAction;
+use App\Actions\Auth\LogoutAction;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
-use App\Http\Requests\ApiFormRequest;
 
 class AuthController extends Controller
 {
-    protected $loginAction;
-    protected $logoutAction;
 
-    public function __construct(LoginAction $loginAction, LogoutAction $logoutAction)
-    {
-        $this->loginAction = $loginAction;
-        $this->logoutAction = $logoutAction;
-    }
+    public function __construct(
+        private LoginAction $loginAction,
+        private LogoutAction $logoutAction){}
 
     // Login
     public function login(Request $request)
